Room Notification Improvements (#43)
* Fix UI test by clicking notification based on text instead of index

* Change notification wording for a better UX

* Move VideoService to room package

* Abstract notification logic outside of the VideoService

* Add chronometer to notification

* Add room name to notification and remove unneeded handler in service
